Initialize task with an empty string
When we declare but do not initialize public $task, the property $task
defaults to null.

Livewire will track $task as part of the component’s state; however,
since it’s null, it might behave differently in Blade template or when
interacting with HTML inputs.

When we bind $task to an input field (e.g., <input type="text"
wire:model="task">), the input field will initially be empty, which is
usually fine.

However, if we rely on $task being a string (e.g., using string
functions like trim() or strlen()), we might encounter issues if $task
is null.