Add Livewire updated hook
The updated method in Livewire is a lifecycle hook that gets called
automatically whenever a public property is updated in the component.
The hook applies to ALL public properties of the component.

Whenever any public property in the Livewire component changes (through
user input, JavaScript, etc.), this updated method gets called.

$property is the name of the property that was updated while $value is
the new value for the property.  The new value is turned into uppercase
and assigned to the property.