Use Livewire $rules property
Define validation rules as $rules property in the Livewire component to
separate them from the save() method.  This keeps the code cleaner and
makes the validation rules reusable.

The use of protected access modifier encapsulates the rules and that
strikes a balance between control and flexibility.

(1) By restricting access to $rules from outside the class hierarchy,
    this prevents unintended modifications or misuse.

(2) By allowing subclasses to access and override $rules, this enables
    customization and reuse of the component.

This balance is particularly important to Livewire, where components are
often extended or reused in different contexts.