Use Alpine.js to build a counter
The x-data directive initializes a reactive state object.  In this case,
count is initialized to 0.  It acts as a local state for this specific
DOM element.

The x-text directive binds the count state to the text content of the
<span> element.  Whenever count changes, the displayed value in the
<span> will automatically update.

The x-on:click directive listens for a click event on the button.  When
the button is clicked, the count state is incremented by 1.